refactor: split GitHub API, unify score normalization, centralize CLI paths and UTC parsing
