+++ gittensor/cli/miner_commands/helpers.py
@@ -5,11 +5,11 @@
 from __future__ import annotations
 
 import json
-from pathlib import Path
 
 import click
 from rich.console import Console
 
+from gittensor.cli.paths import load_config_value
 from gittensor.constants import NETWORK_MAP
 
 console = Console()
@@ -28,26 +28,14 @@ def _get_validator_axons(metagraph) -> tuple[list, list]:
     return axons, uids
 
 
-def _load_config_value(key: str):
-    """Load a value from ~/.gittensor/config.json, or None."""
-    config_file = Path.home() / '.gittensor' / 'config.json'
-    if not config_file.exists():
-        return None
-    try:
-        config = json.loads(config_file.read_text())
-        return config.get(key)
-    except (json.JSONDecodeError, OSError):
-        return None
-
-
 def _resolve_endpoint(network: str | None, rpc_url: str | None) -> str:
     """Resolve the subtensor endpoint from CLI args or config."""
     if rpc_url:
         return rpc_url
     if network:
         return NETWORK_MAP.get(network.lower(), network)
-    config_network = _load_config_value('network')
-    config_endpoint = _load_config_value('ws_endpoint')
+    config_network = load_config_value('network')
+    config_endpoint = load_config_value('ws_endpoint')
     if config_endpoint:
         return config_endpoint
     if config_network:
