+++ gittensor/validator/oss_contributions/normalize.py
@@ -1,35 +1,17 @@
 from typing import Dict
 
-import bittensor as bt
-
 from gittensor.classes import MinerEvaluation
+from gittensor.validator.score_normalization import normalize_miner_evaluations_linear
 
 
 def normalize_rewards_linear(miner_evaluations: Dict[int, MinerEvaluation]) -> Dict[int, float]:
     """Normalize scores to sum to 1.0, preserving ratios."""
 
-    if not miner_evaluations:
-        bt.logging.warning('No miner evaluations provided for normalization')
-        return {}
-
-    rewards: Dict[int, float] = {}
-    zero_reward_count = 0
-
-    for uid, evaluation in miner_evaluations.items():
-        rewards[uid] = evaluation.total_score
-        if rewards[uid] > 0:
-            bt.logging.info(f'Final reward for uid {uid}: {rewards[uid]:.2f}')
-        else:
-            zero_reward_count += 1
-
-    if zero_reward_count > 0:
-        bt.logging.info(f'{zero_reward_count} miners have 0 reward')
-
-    total = sum(rewards.values())
-    if total <= 0:
-        bt.logging.info('All scores are zero, returning original scores')
-        return rewards
-
-    normalized = {uid: score / total for uid, score in rewards.items()}
-
-    return normalized
+    return normalize_miner_evaluations_linear(
+        miner_evaluations,
+        lambda e: e.total_score,
+        empty_warning='No miner evaluations provided for normalization',
+        log_each_positive_uid=True,
+        count_zero_log_label='0 reward',
+        all_zero_log='All scores are zero, returning original scores',
+    )
